requirements.txt
streamlit
PyPDF2
sentence-transformers
faiss-cpu
huggingface-hub
python-dotenv
